V3 rajout documentation pour doxygen + un peu de néttoyage

+++ function.c
@@ -14,7 +14,16 @@
 #include "function.h"
 #include <string.h>
 
-
+/**
+ * \fn void stateRegister(int* reg, int PC, int SP, int SR)
+ * \brief Affiche l'état des registre, utilisé pour tester avant l'utilisation de ncurses.
+ *
+ * \param reg Tableau contenant les registres.
+ * \param PC registre PC
+ * \param SP registre SP
+ * \param SR registre SR
+ * \return EXIT_SUCCESS - Arrêt normal du programme.
+ */
 void stateRegister(int* reg, int PC, int SP, int SR){
         printf("\n\nEtat des registres :");
 
@@ -29,6 +38,13 @@ void stateRegister(int* reg, int PC, int SP, int SR){
         printf("\nSR : 0x%s",toHexa(SR));
 }
 
+/**
+ * \fn char* toHexa(int n)
+ * \brief Convertie un entier en Hexadecimal (sous forme de String)
+ *
+ * \param n l'entier à convertire.
+ * \return Une chaine contenant l'entier donnée convertie en hexa (inutile puisque sprintf permet déjà de le faire --> a supprimer dans une version future)
+ */
 char* toHexa(int n){
 
     char* retour = malloc(sizeof(char)*40);
@@ -61,6 +77,13 @@ char* toHexa(int n){
     return retour;
 }
 
+/**
+ * \fn char* toBinaire(int n)
+ * \brief Convertie un entier en binaire (sous forme de String)
+ *
+ * \param n l'entier à convertire.
+ * \return Une chaine contenant l'entier donnée convertie en binaire (inutile puisque sprintf permet déjà de le faire --> a supprimer dans une version future)
+ */
 char* toBinaire(int n){
 
     char* retour = malloc(sizeof(char)*40);
@@ -79,6 +102,14 @@ char* toBinaire(int n){
     return retour;
 }
 
+/**
+ * \fn char* add0(char* bin, int size)
+ * \brief Ajoute des 0 en tête d'une chaine
+ *
+ * \param bin Chaine contenant un nombre binaire
+ * \param taille de la chaine finale (après rajout des 0)
+ * \return Une chaine contenant un binaire de taille size (inutile puisque sprintf permet déjà de le faire --> a supprimer dans une version future)
+ */
 char* add0(char* bin, int size){
 
     // Chaine finale comportant tous les 0 en tête du nombre binaire
@@ -104,13 +135,29 @@ char* add0(char* bin, int size){
     return ch0;
 }
 
-void add(short int* source, int* dest, int* SR){
+/**
+ * \fn void add(int* source, int* dest, int* SR)
+ * \brief Fonction add de l'émulateur ASM (dest = dest + source).
+ *
+ * \param source Valeur de l'opérande source.
+ * \param dest Valeur de l'opérande de destination
+ * \param adresse du Status Register
+ */
+void add( int* source, int* dest, int* SR){
     *dest = *dest + *source;
     // On retourne le résultat pour le stocker dans le registre SR
     *SR = *dest;
 }
 
-void sub(short int* source, int* dest, int* SR){
+/**
+ * \fn void sub(int* source, int* dest, int* SR)
+ * \brief Fonction sub de l'émulateur ASM (dest = dest - source).
+ *
+ * \param source Valeur de l'opérande source.
+ * \param dest Valeur de l'opérande de destination
+ * \param adresse du Status Register
+ */
+void sub( int* source, int* dest, int* SR){
     *dest = *dest - *source;
     // On retourne le résultat pour le stocker dans le registre SR
     *SR = *dest;
